+++ lib/nginx_proxy.sh
@@ -21,25 +21,18 @@
 
 # Reloads Nginx
 function nginx_reload() {
-    local _jid="${1}"
-    local _exitCode
     service nginx reload > /dev/null 2>&1
     
-    _exitCode=$?
-    
-    if [[ ${_exitCode} -ne 0 ]]; then
-        e_error "Failed to reload HTTP proxy"
-    else
-         e_verbose "Reloaded HTTP proxy" 
-    fi
-    return ${_exitCode}
+    return $?
 }
 
 # Creates a server file for nginx
 function nginx_create_servername_file() {
+
     local _subdomain="${1}"
     local _filePath="${2}"
     local _certificate="${3}"
+
     local _certificateLine=""
     
     # make sure the certificate exists or nginx will error
@@ -111,6 +104,7 @@ function nginx_add_location_block() {
 
     # check if this definition already exists in the nginx config
     if [[ $(cat "${_filePath}" | grep "location ${_urlPath} {" | wc -l ) -eq 0 ]]; then
+
         # check if its an ssl location or not
         if [[ "${_ssl}" == "true" ]]; then
             _listenLine="${_listenLine}:443 ssl;"
@@ -124,7 +118,7 @@ function nginx_add_location_block() {
         $(add_line_to_file_after_string "    location ${_urlPath} {" "${_listenLine}" "${_filePath}")
         # add the closing brace
         $(add_line_to_file_after_string "    }" "    location ${_urlPath} {" "${_filePath}")
-
+        
         return ${E_SUCCESS}
     else
         e_verbose "Location data already in ${_filePath}"
@@ -157,10 +151,11 @@ function nginx_insert_location_data() {
 
     # check if this url is a websocket url and add in the relevant config
     if [[ "${_urlWebSocket}" == "yes" ]]; then
-        add_line_to_file_between_strings_if_not_exists "location ${_urlDirectory} {" "        proxy_http_version 1.1;" "}" "${_filePath}"
-        add_line_to_file_between_strings_if_not_exists "location ${_urlDirectory} {" '        proxy_set_header Upgrade $http_upgrade;' "}" "${_filePath}"
-        add_line_to_file_between_strings_if_not_exists "location ${_urlDirectory} {" '        proxy_set_header Connection "upgrade";' "}" "${_filePath}"
-        add_line_to_file_between_strings_if_not_exists "location ${_urlDirectory} {" "        proxy_read_timeout 600;" "}" "${_filePath}"
+        # include the websockets proxy file
+        add_line_to_file_between_strings_if_not_exists "location ${_urlDirectory} {" "        include proxy_pass/ws_wss;" "}" "${_filePath}"
+    else
+        # include the standard HTTP(S) proxy file
+        add_line_to_file_between_strings_if_not_exists "location ${_urlDirectory} {" "        include proxy_pass/http_https;" "}" "${_filePath}"
     fi
 
     # check if this url has a max file size setting
@@ -209,20 +204,35 @@ function nginx_create_access_file() {
         fi
     } > "${_accessFile}"
     
-    return $E_SUCCESS
+    return ${E_SUCCESS}
 }
 
+# clears out an access file
+function nginx_clear_access_file() {
+    local _accessFile="${1}"
+    
+    if [[ -f "${_accessFile}" ]]; then
+        echo '' > "${_accessFile}"
+        
+        return $?
+    fi
+    
+    
+    return ${E_ERROR}
+}
 
 # formats a given filename into the correct format for nginx
 function nginx_format_filename() {
     local filename="${1}"
+    # swap :// for dash
+    filename=$(echo "${filename}" | sed "s|://|-|" )
     # swap dots for underscores
     filename=$(echo "${filename}" | tr '.' '_')
     # and slashes for dashes
     filename=$(echo "${filename}" | tr '/' '-')
     
     echo "${filename}"
-    return $E_SUCCESS
+    return ${E_SUCCESS}
     
 }
 
@@ -233,10 +243,10 @@ function nginx_remove_include() {
     
     # remove the lines from the file
     if remove_lines_from_file "${_file}" "include ${_include};" "false"; then
-        return $E_SUCCESS
+        return ${E_SUCCESS}
     fi
     
-    return $E_ERROR
+    return ${E_ERROR}
 }
 
 # sets up a url with given parameters
@@ -249,6 +259,8 @@ function nginx_add_url() {
     local _uuid="${6}"
     local _container_dataset="${7}"
     declare -a _whiteList=("${!8}")
+    
+    local _urlDomain _urlDirectory _filename _upstreamFilename
     
     # split up the url into its domain and directory segments
     # check if the url actually contained a /
@@ -276,11 +288,10 @@ function nginx_add_url() {
         #####################################
         # SET UP THE HTTPS UPSTREAM FILE
         # check if the https upstream file exists
-        if [[ ! -f "${NGINX_UPSTREAM_DIR}/https-${_filename}" ]]; then
+        if [[ ! -f "${NGINX_UPSTREAM_DIR}/https-${_upstreamFilename}" ]]; then
             # create it
-    
             if ! nginx_create_upstream_file "https-${_upstreamFilename}" "${NGINX_UPSTREAM_DIR}/https-${_upstreamFilename}"; then
-                e_error "Failed to create HTTP proxy upstream file ${NGINX_UPSTREAM_DIR}/https-${_filename}"
+                e_error "Failed to create HTTP proxy upstream file ${NGINX_UPSTREAM_DIR}/https-${_upstreamFilename}"
             fi
         fi
         
@@ -309,20 +320,10 @@ function nginx_add_url() {
         
         #####################################
         # SET UP THE HTTP REDIRECT SERVER_NAME FILE
-        # check if the https server_name file exists
-        if [[ ! -f "${NGINX_SERVERNAME_DIR}/http-${_filename}" ]]; then
-            # create it
-            if ! nginx_create_servername_file "${_urlDomain}" "${NGINX_SERVERNAME_DIR}/http-${_filename}" ""; then
-                e_error "Failed to create HTTP proxy servername file ${NGINX_SERVERNAME_DIR}/http-${_filename}"
-            fi
-        fi
-        
-        # add the location data if it doesnt already exist
-        nginx_add_location_block "${_urlDirectory}" "${NGINX_SERVERNAME_DIR}/http-${_filename}" "false"
-        # insert the redirect
-        $(add_line_to_file_between_strings_if_not_exists "location ${_urlDirectory} {" '        return 301 https://$host$request_uri;' "}" "${NGINX_SERVERNAME_DIR}/http-${_filename}")
-        # include this file in the dataset for destruction
-        zfs_append_custom_array "${_container_dataset}" "${ZFS_PROP_ROOT}.nginx_servername" "http-${_filename}"
+        nginx_add_redirect_url "http://${_url}" "https://${_url}"
+
+        # register the redirect url within zfs
+        zfs_append_custom_array "${_container_dataset}" "${ZFS_PROP_ROOT}.redirect_url" "http://${_url}"
         
         # Include the access file for this container in the server_name file
         local _accessFileName=$( nginx_format_filename "${_uuid}" )
@@ -334,7 +335,7 @@ function nginx_add_url() {
             nginx_create_access_file "${_accessFilePath}" _whiteList[@] "true"
             # add a deny all into the server file so that we can reference many allows from includes
             #$(add_line_to_file_after_string "        deny all;" "location ${_urlDirectory} {" "${NGINX_SERVERNAME_DIR}/https-${_filename}")
-        
+            
             # and include the access file for this container
             $(add_line_to_file_after_string "        include ${_accessFilePath};" "location ${_urlDirectory} {" "${NGINX_SERVERNAME_DIR}/https-${_filename}");
         fi
@@ -342,10 +343,10 @@ function nginx_add_url() {
         #####################################
         # SET UP THE HTTP UPSTREAM FILE
         # check if the https upstream file exists
-        if [[ ! -f "${NGINX_UPSTREAM_DIR}/http-${_filename}" ]]; then
+        if [[ ! -f "${NGINX_UPSTREAM_DIR}/http-${_upstreamFilename}" ]]; then
             # create it
             if ! nginx_create_upstream_file "http-${_upstreamFilename}" "${NGINX_UPSTREAM_DIR}/http-${_upstreamFilename}"; then
-                e_error "Failed to create HTTP proxy upstream file ${NGINX_UPSTREAM_DIR}/http-${_filename}"
+                e_error "Failed to create HTTP proxy upstream file ${NGINX_UPSTREAM_DIR}/http-${_upstreamFilename}"
             fi
         fi
         
@@ -353,7 +354,7 @@ function nginx_add_url() {
         nginx_add_to_upstream_block "${NGINX_UPSTREAM_DIR}/http-${_upstreamFilename}" "${_ip4}" "80" "http-${_upstreamFilename}"
         # include this file in the dataset for destruction
         zfs_append_custom_array "${_container_dataset}" "${ZFS_PROP_ROOT}.nginx_upstream" "http-${_upstreamFilename}"
-        
+        set +x
         #####################################
         # SET UP THE HTTP SERVER_NAME FILE
 
@@ -386,4 +387,77 @@ function nginx_add_url() {
             $(add_line_to_file_after_string "        include ${_accessFilePath};" "location ${_urlDirectory} {" "${NGINX_SERVERNAME_DIR}/http-${_filename}");
         fi
     fi
+}
+
+# adds a url definition that is redirected to another url
+# takes a single from and a single to
+function nginx_add_redirect_url() {
+    local _redirectFrom="${1}"
+    local _redirectTo="${2}"
+    local _redirectFromCert="${3}"
+    
+    local _urlFromCert _urlFromDomain _urlFromDirectory _transformedRedirectFrom
+
+    # check if it starts with a protocol
+    if [[ "${_redirectFrom}" =~ ^https ]]; then
+        # HTTPS url
+        _urlFromSSL="true"
+        # strip off the https part
+        _transformedRedirectFrom=${_redirectFrom#https://}
+        
+    elif [[ "${_redirectFrom}" =~ ^http ]]; then
+        _urlFromSSL="false"
+        # strip off the http part
+        _transformedRedirectFrom=${_redirectFrom#http://}
+    else
+        # no protocol received so error
+        return ${E_ERROR}
+    fi
+    
+    # check if the url actually contained a /
+    if string_contains_char "${_transformedRedirectFrom}" '/'; then
+        _urlFromDomain=$(lcut ${_transformedRedirectFrom} '/')
+        _urlFromDirectory=$(rcut ${_transformedRedirectFrom} '/')
+        # add the / back in
+        _urlFromDirectory="/${_urlFromDirectory}"
+    else
+        _urlFromDomain="${_transformedRedirectFrom}"
+        _urlFromDirectory='/'
+    fi
+    
+    # remove any trailing slashes
+    local _servernameFilename=$(rtrim "${_urlFromDomain}" '/')
+    local _upstreamFilename=$( rtrim "${_transformedRedirectFrom}" '/' )
+
+    # format the filename of the file to edit - swap dots for underscores
+    _servernameFilename=$( nginx_format_filename "${_servernameFilename}" )
+    _upstreamFilename=$( nginx_format_filename "${_upstreamFilename}" )
+    
+    # prepend the protocol to the start of servername filename
+    if [[ "${_redirectFrom}" =~ ^https ]]; then
+        _servernameFilename="https-${_servernameFilename}"
+        _urlFromCert="${_redirectFromCert}"
+    else 
+        _servernameFilename="http-${_servernameFilename}"
+    fi
+    
+    # check if the https server_name file exists
+    if [[ ! -f "${NGINX_SERVERNAME_DIR}/${_servernameFilename}" ]]; then
+        # create it
+        if ! nginx_create_servername_file "${_urlFromDomain}" "${NGINX_SERVERNAME_DIR}/${_servernameFilename}" "${_urlFromCert}"; then
+            return ${E_ERROR}
+        fi
+    fi
+
+    # add the location data if it doesnt already exist
+    if ! nginx_add_location_block "${_urlFromDirectory}" "${NGINX_SERVERNAME_DIR}/${_servernameFilename}" "${_urlFromSSL}"; then
+        return ${E_ERROR}
+    fi
+
+    # insert the redirect
+    if ! add_line_to_file_between_strings_if_not_exists "location ${_urlFromDirectory} {" "        return 301 ${_redirectTo}\$request_uri;" "}" "${NGINX_SERVERNAME_DIR}/${_servernameFilename}"; then
+        return ${E_ERROR}
+    fi
+    
+    return ${E_SUCCESS}
 }