+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.10.0 [Unreleased]
+#### Added
+- Added this CHANGELOG.md file
+- Added CONTRIBUTING.md guidelines for contribution
+- Implemented urlRedirect= for URLs
+- Implemented ability to change container resource limits/whitelist
+  on the fly
+
 ## 0.9.0 - 2016-04-21
 #### Added
 - Initial release of Tredly
+
+[Unreleased]: https://github.com/tredly/tredly-build/compare/0.9.0...HEAD